styling custom alerts and confirming delete patient

+++ src/account.js
@@ -2,10 +2,14 @@ import React, { useEffect, useState } from 'react';
 import './account.css';
 import { useNavigate } from 'react-router-dom';
 import { FaEdit, FaTrash } from 'react-icons/fa';
+import Alert from './components/Alert';
 
 const Account = () => {
   const [userList, setUserList] = useState([]);
   const navigate = useNavigate();
+
+  const [alert, setAlert] = useState(null);
+  const [showAlert, setShowAlert] = useState(true);
 
   useEffect(() => {
     const storedList = localStorage.getItem("userInfoList");
@@ -14,11 +18,20 @@ const Account = () => {
     }
   }, []);
 
+  
   const handleDelete = (id) => {
-    const updatedList = userList.filter(user => user.id !== id);
-    setUserList(updatedList);
-    localStorage.setItem("userInfoList", JSON.stringify(updatedList));
+    const confirmDelete = window.confirm("Are you sure you want to delete this patient?");
+    if (confirmDelete) {
+      setAlert({ message: "This patient will be deleted", type: "warning" });
+      setTimeout(() => {
+        const updatedList = userList.filter(user => user.id !== id);
+        setUserList(updatedList);
+        localStorage.setItem("userInfoList", JSON.stringify(updatedList));
+        setAlert(null);
+      }, 2000);
+    }
   };
+
 
   const handleEdit = (user) => {
     localStorage.setItem("editPatient", JSON.stringify(user));
@@ -36,6 +49,13 @@ const Account = () => {
 
   return (
     <div className="AccountInfo">
+      {alert && (
+        <Alert
+          message={alert.message}
+          type={alert.type}
+          onClose={() => setAlert(null)}
+        />
+      )}
       <h1>My Patient Information</h1>
       <a href="info" className="edit">Add Patients</a>
       {userList.map((user) => (
